+++ complex_rest/core/urls.py
@@ -19,10 +19,15 @@ from django.contrib import admin
 from django.urls import path, include
 from core.load_plugins import get_plugin_base_url
 
+from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
+
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('auth/', include(('rest_auth.urls', 'rest_auth'), namespace='auth')),
+    path('schema/', SpectacularAPIView.as_view(), name='schema'),
+    path('schema/swagger-ui/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
+    path('schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
 ]
 
 # add plugins urls if they exist except plugin examples when not debug