+++ docs/deploy/collectstatic.sh
@@ -1,0 +1,5 @@
+cd `dirname "${BASH_SOURCE[0]}"`
+
+source ./venv/bin/activate
+
+python ./complex_rest/manage.py collectstatic